+++ src/components/Target.jsx
@@ -1,0 +1,6 @@
+// 인피니티 스크롤 적용시 사용
+const Target = ({ ref }) => {
+  return <div className='h-[1px]' ref={ref}></div>;
+};
+
+export default Target;
