+++ src/pages/Products.jsx
@@ -3,40 +3,18 @@ import { DEFAULT_META_DATA_URL } from '@constants/url';
 import { useParams } from 'react-router';
 import ProductCardContainer from '@components/ui/common/ProductCardContainer';
 import getProductsByCategory from '@api/getProductsByCategory';
-import { useEffect, useState } from 'react';
-import usePagination from '@hooks/usePagination';
-import useIntersect from '@hooks/useIntersect';
-import useFetch from '@hooks/useFetch';
+import useProductInfinityScroll from '@hooks/useProductInfinityScroll';
+import Target from '@components/Target';
 
 const Products = () => {
-  const [products, setProducts] = useState([]);
-  const { catalog } = useParams();
   const countPerPage = 6;
+  const { catalog } = useParams();
 
-  const { fetchNextPageQuery, hasNextPage, setHasNextPage, setCurrentSkip } = usePagination({
+  const { products, total, isLoading, error, ref } = useProductInfinityScroll({
+    countPerPage,
     query: getProductsByCategory,
     options: { categoryName: catalog },
-    countPerPage,
   });
-
-  const { data, isLoading, error } = useFetch({ query: fetchNextPageQuery });
-
-  const total = data?.total ?? 0;
-  const ref = useIntersect({
-    onIntersect: async (entry, observer) => {
-      observer.unobserve(entry.target);
-      if (!isLoading && hasNextPage) {
-        setCurrentSkip((prev) => prev + countPerPage);
-      }
-    },
-  });
-
-  useEffect(() => {
-    if (data && data.products) {
-      setProducts((prevProducts) => [...prevProducts, ...data.products]);
-      setHasNextPage(data.products.length === countPerPage);
-    }
-  }, [data, setHasNextPage]);
 
   if (error) {
     console.log(error);
@@ -55,16 +33,16 @@ const Products = () => {
       <div className='flex justify-center'>
         <div className='flex flex-col items-start'>
           <div>Products Result : {total}</div>
-          <ProductCardContainer products={products} isLoading={isLoading} skeletonSize={6} />
+          <ProductCardContainer
+            products={products}
+            isLoading={isLoading}
+            skeletonSize={countPerPage}
+          />
           <Target ref={ref} />
         </div>
       </div>
     </div>
   );
 };
-
-function Target({ ref }) {
-  return <div className='h-[1px]' ref={ref}></div>;
-}
 
 export default Products;
